devtools: move stuff out of utils.py that has only one user
