+++ dev/releases/make_github_release.py
@@ -10,6 +10,8 @@
 ##
 ##  This script makes a github release and uploads all tarballs as assets.
 ##
+import re
+import subprocess
 import sys
 
 import utils
@@ -19,12 +21,51 @@ from utils import error, notice
 if len(sys.argv) != 3:
     error("usage: " + sys.argv[0] + " <tag_name> <path_to_release>")
 
+
+def is_possible_gap_release_tag(tag: str) -> bool:
+    return re.fullmatch(r"v[1-9]+\.[0-9]+\.[0-9]+(-.+)?", tag) is not None
+
+
+def verify_is_possible_gap_release_tag(tag: str) -> None:
+    if not is_possible_gap_release_tag(tag):
+        error(f"{tag} does not look like the tag of a GAP release version")
+
+
+# lightweight vs annotated
+# https://stackoverflow.com/questions/40479712/how-can-i-tell-if-a-given-git-tag-is-annotated-or-lightweight#40499437
+def is_annotated_git_tag(tag: str) -> bool:
+    res = subprocess.run(
+        ["git", "for-each-ref", "refs/tags/" + tag],
+        capture_output=True,
+        text=True,
+        check=False,
+    )
+    return res.returncode == 0 and res.stdout.split()[1] == "tag"
+
+
+def check_git_tag_for_release(tag: str) -> None:
+    if not is_annotated_git_tag(tag):
+        error(f"There is no annotated tag {tag}")
+    # check that tag points to HEAD
+    tag_commit = subprocess.run(
+        ["git", "rev-parse", tag + "^{}"], check=True, capture_output=True, text=True
+    ).stdout.strip()
+    head = subprocess.run(
+        ["git", "rev-parse", "HEAD"], check=True, capture_output=True, text=True
+    ).stdout.strip()
+    if tag_commit != head:
+        error(
+            f"The tag {tag} does not point to the current commit {head} but"
+            + f" instead points to {tag_commit}"
+        )
+
+
 TAG_NAME = sys.argv[1]
 PATH_TO_RELEASE = sys.argv[2]
 VERSION = TAG_NAME[1:]  # strip 'v' prefix
 
 utils.verify_git_clean()
-utils.verify_is_possible_gap_release_tag(TAG_NAME)
+verify_is_possible_gap_release_tag(TAG_NAME)
 repo = utils_github.initialize_github()
 
 # Error if the tag TAG_NAME hasn't been pushed out yet.
@@ -35,7 +76,7 @@ if not any(tag.name == TAG_NAME for tag in repo.get_tags()):
 # - exists
 # - is an annotated tag
 # - points to current HEAD
-utils.check_git_tag_for_release(TAG_NAME)
+check_git_tag_for_release(TAG_NAME)
 
 # Error if this release has been already created on GitHub
 if any(r.tag_name == TAG_NAME for r in repo.get_releases()):
